Reconnect changes
